Compare unsigned/unsigned instead of unsigned/signed
Monotone-Parent: 9cd235487bac82f9a54a17fdc4972905195754d6
Monotone-Revision: dea797f2f9523d487ece1c34f8172e9e9d193a85

+++ test/test_recovery.cpp
@@ -129,7 +129,7 @@ bool check_recovery(byte k, byte n, const std::string& hex_input,
          {
          //printf("%02X", pkts[i][j]);
 
-         if(pkts[i][j] != input[stride*i+j])
+         if(pkts[i][j] != (byte)input[stride*i+j])
             printf("Bad: pkts[%d][%d] = %02X != input[%d*%d+%d] = %02X\n",
                    i, j, pkts[i][j], stride, i, j, (byte)input[stride*i+j]);
          }
